+++ packages/clients/CHANGELOG.md
@@ -8,10 +8,10 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-* **clients:** incorrect import ([10b3b0b](https://github.com/marcelovicentegc/octosync/commit/10b3b0bac2e4a6265e20d0fbfae36341b488358e))
+* **clients:** incorrect import ([10b3b0b](https://github.com/howl0893/IssueForge/commit/10b3b0bac2e4a6265e20d0fbfae36341b488358e))
 
 
 ### Features
 
-* **utils:** move useEnv to utils package ([87be656](https://github.com/marcelovicentegc/octosync/commit/87be65669310125d6733a4adeb229d18be6dd7ea))
-* split packages to include octosync-bot [skip-deploy] ([36a5565](https://github.com/marcelovicentegc/octosync/commit/36a55659ea20b5ab7d628f0b207a0c8400c86d43))
+* **utils:** move useEnv to utils package ([87be656](https://github.com/howl0893/IssueForge/commit/87be65669310125d6733a4adeb229d18be6dd7ea))
+* split packages to include IssueForge-bot [skip-deploy] ([36a5565](https://github.com/howl0893/IssueForge/commit/36a55659ea20b5ab7d628f0b207a0c8400c86d43))